NelderMead.jl Documentation

NelderMead.optimise! — Method

optimise!(s, f; kwargs...)

Find minimum of function, f, starting from Simplex, s, with options passed in via kwargs.

Keyword Arguments

stopval (default -Inf): stopping criterion when function evaluates

equal to or less than stopval

xtol_abs (default zeros(T)) .* ones(Bool, dimensionality(s)): stop if

the vertices of simplex get within this absolute tolerance

xtol_rel (default eps(T)) .* ones(Bool, dimensionality(s)): stop if

the vertices of simplex get within this relative tolerance

ftol_abs (default zero(real(U))): stop if function evaluations at the

vertices are close to one another by this absolute tolerance

ftol_rel (default 1000eps(real(U))): stop if function evaluations at the

vertices are close to one another by this relative tolerance

maxiters (default 1000): maximum number of iterations of the Nelder Mead

algorithm

timelimit (default Inf): stop if it takes longer than this in seconds

α (default 1): Reflection factor

β (default 0.5): Contraction factor

γ (default 2): Expansion factor

δ (default 0.5): Shrinkage factor

Returns

optimise returns a tuple consisting of:

minimiser: the location of the minimum

minimumvalue: the value at the minimum

returncode: the return code symbol

numiters: the number of iterations taken

simplex: the simplex in its final state (useful for restarting)

NelderMead.optimise — Method

optimise(f, initial_positions, initial_step; kwargs...)

Find minimum of function, f, first creating a Simplex using a starting vertex position, initial_position, and other vertices initial_step away from that point in all directions, and options passed in via kwargs.

NelderMead.optimise — Method

optimise(f, initial_vertex_positions; kwargs...)

Find minimum of function, f, first creating a Simplex from vertices at initial_vertex_positions, and options passed in via kwargs.

Index

NelderMead.optimise

NelderMead.optimise

NelderMead.optimise!